+++ src/data/store/key/map/define.rs
@@ -23,7 +23,7 @@ map! {
 #[doc = crate::_tags!(construction data_structure)]
 /// Defines a custom static hashmap.
 #[doc = crate::_doc_meta!{
-    location("data/id", macro map),
+    location("data/store/key", macro map),
 }]
 /// # Arguments
 /// - `$NAME`:      the name of the new hashmap.